requirements.txt
requests==2.31.0
pandas==2.1.4
kafka-python==2.0.2
minio==7.1.15
websocket-client==1.6.1
yfinance==0.2.55
ipython==8.14.0
websocket-client==1.6.1